inspect moved
