+++ x-pack/elastic-agent/pkg/agent/cmd/enroll_cmd_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package application
+package cmd
 
 import (
 	"bytes"
@@ -83,7 +83,7 @@ func TestEnroll(t *testing.T) {
 			store := &mockStore{Err: errors.New("fail to save")}
 			cmd, err := NewEnrollCmdWithStore(
 				log,
-				&EnrollCmdOption{
+				&enrollCmdOption{
 					ID:                   "my-id",
 					URL:                  url,
 					CAs:                  []string{caFile},
@@ -136,7 +136,7 @@ func TestEnroll(t *testing.T) {
 			store := &mockStore{}
 			cmd, err := NewEnrollCmdWithStore(
 				log,
-				&EnrollCmdOption{
+				&enrollCmdOption{
 					ID:                   "my-id",
 					URL:                  url,
 					CAs:                  []string{caFile},
@@ -193,7 +193,7 @@ func TestEnroll(t *testing.T) {
 			store := &mockStore{}
 			cmd, err := NewEnrollCmdWithStore(
 				log,
-				&EnrollCmdOption{
+				&enrollCmdOption{
 					ID:                   "my-id",
 					URL:                  url,
 					CAs:                  []string{},
@@ -253,7 +253,7 @@ func TestEnroll(t *testing.T) {
 			store := &mockStore{}
 			cmd, err := NewEnrollCmdWithStore(
 				log,
-				&EnrollCmdOption{
+				&enrollCmdOption{
 					ID:                   "my-id",
 					URL:                  url,
 					CAs:                  []string{},
@@ -298,7 +298,7 @@ func TestEnroll(t *testing.T) {
 			store := &mockStore{}
 			cmd, err := NewEnrollCmdWithStore(
 				log,
-				&EnrollCmdOption{
+				&enrollCmdOption{
 					ID:                   "my-id",
 					URL:                  url,
 					CAs:                  []string{},
